feat: factory method working on netbeans 17

+++ FactoryMethod/src/main/java/entitys/Product.java
@@ -1,25 +1,30 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
+ */
 package entitys;
 
 /**
  *
- * @author Oscar Blancarte <oscarblancarte3@gmail.com>
+ * @author richa
  */
 public class Product {
-    private Long id;
+    private int id;
     private String name;
     private double price;
 
-    public Product(Long id, String name, double price) {
+    public Product(int id, String name, double price) {
         this.id = id;
         this.name = name;
         this.price = price;
     }
     
-    public Long getId() {
+    
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
 
@@ -43,4 +48,7 @@ public class Product {
     public String toString() {
         return "Product{" + "id=" + id + ", name=" + name + ", price=" + price + '}';
     }
+    
+    
+    
 }
